requirements.txt
python-telegram-bot==21.5
qrcode==8.0
pillow==11.0.0
